Backend/requirements_new.txt
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database & ORM
SQLAlchemy==2.0.23
Alembic==1.12.0

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0

# Logging & Monitoring
sentry-sdk[flask]==1.38.0

# API Documentation
flask-swagger-ui==4.11.1
flask-restx==1.3.0

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Development Tools
black==23.7.0
flake8==6.0.0
